requirements.txt
playwright>=1.40.0
pytest>=7.0.0
pytest-html>=4.0.0
pyyaml>=6.0
allure-pytest>=2.9.0
